[Library] Add  `locatableContent` traits/specifications
Within the Media Creation industry, entities often hold data in addition
to the fields communicated through the OpenAssetIO API.

A common example of this is an image file, where the actual pixel data
is held in a file on disk, or accessible through some other URL.

The strategy for representing this in the API is documented here:

https://github.com/TheFoundryVisionmongers/OpenAssetIO/blob/main/decisions/decisions/DR009-Locatable-content-traits.md

This commit adds prototype traits and specifications representing
Option 3 of the document.

There is one notable change to the sketch proposed there. During
implementation, it became apparent that the "default" value for the
`usesURLs` property of the `locatableContentReader` trait as sketched
is ambiguous, as is the fact that `supoortedURLSchemes` property only
applies when `usesURLs` is set to `True`.

This is somewhat of an anti-pattern, as OpenAssetIO strives towards
explicit over implicit wherever possible.

In this implementation, we rename `locatableContentReader` to simply
`usesUrls`. This has the following benefits:

 - The trait is set, or not, and when set, defines a specific behaviour.
 - Any additional properties of that trait are then clearly only
   associated with the behaviour imbued by the trait.

We also remove the `supportedURLSchemes` property until suitable
data types are refined for trait properties - using a map + bool creates
ambiguity as to how it should be interpreted.

Signed-off-by: Tom Cowland <tom@foundry.com>