[Library] Make `locatableContent` URL only
This a a minefield. See the options and discussion here:

  https://github.com/TheFoundryVisionmongers/OpenAssetIO/pull/356

The short of it is that anything that dynamically changes the _format_
of a trait property value based on the locale breaks the ability for a
manager to generically store and recall trait data - as its
interpretation is no longer constant. Being able to persist generic
entity data is (at present) a core API requirement.

Using the context to rewrite the _value_ of a trait property does not
suffer from the same issues as the interpretation of the data is always
constant.

For now, we revert to the simplest (logically speaking) approach, even
though the string manipulation/encoding required may add significant
overhead.

We can pick this debate up when this is actually surfaced as an issue,
and we have enough like-minded people from the broader community with
their head in the problem space to come up with an acceptable
compromise. We don't have the breadth of experience to make this call
yet.

Signed-off-by: Tom Cowland <tom@foundry.com>