[Traits] Add `ResolvesFutureEntitiesTrait`
Adds the `ResolvesFutureEntities` management policy trait, that can be
used to indicate that a manager is capable of resolving the specified
trait set for entities that don't exist yet. This supersedes the old
`WillManagePathTrait`, that was deprecated and subsequently remove
from OpenAssetIO core API following the switch from the old 'primary
string and attributes' approach to composed traits.

Path centric operations don't make sense in the traits world. There was
some debate as to whether this needs to be finer-grained, and paired
with specific entity traits, but given the managementPolicy call is
specific to a given trait set, it didn't seem necessary. This isn't
suggesting that the manager can resolve any traits, just those queried.

Signed-off-by: Tom Cowland <tom@foundry.com>