[Traits] Add DisplayName trait in `identity` namespace
OpenAssetIO used to have `entityName` and `entityDisplayName` methods.
It makes more sense to use a trait for entity identity as the specifics
of this may well be industry specific, and it potentially avoids several
additional API calls in common scenarios.

The definitions are largely lifed from the old API methods, with minor
massaging of property names to attempt to be more coherent.

Part of #OpenAssetIO/837

Signed-off-by: Tom Cowland <tom@foundry.com>