[Traits] Remove speculative timeline traits
These had originally been added for the `otio-openassetio` prototype.
This has since stopped using them on the basis that highly specific
locales tend to preclude batching (adding data that is specific to any
given reference to the locale means you can't batch the queries as the
context is shared with all references in the batch). Until all this
settles down, and we have a significant number of real integrations
producing scenarios where we do need more information, we should aim to
keep the locale as minimal as possible to maximize batching
opportunities.

Signed-off-by: Tom Cowland <tom@foundry.com>