[Test] Simplify C++ header imports
We previously explicitly imported all the C++ headers. This served to
verify that the YAML produced the correct outputs. We also do this in
Python though, so this is really mostly testing TraitGen.

We have a rule of thumb that we don't re-test auto-generated code, and
the C++ tooling requirements have been noted to form a barrier to entry
to contributors who may work mostly in the Python world.

As, assuming we trust TraitGen, all generated headers are still imported
by the package header, this commit removes the explicit imports of
individual trait headers. This means we should still test that all
shipping headers are syntax correct.

Signed-off-by: Tom Cowland <tom@foundry.com>