[Traits] Add versioning traits/specifications
Adds traits and specifications that can be used to replicate the old
"versioning" API methods from the core API. These were recently removed
in favour of domain specific definitions such as these.

We opted to add base traits for 'usage' to the package. These should be
composed with the other relevant traits when defining a specification to
allow runtime introspection of the trait set. Amongst other uses, this
helps managers in their implementation of methods such as
`managementPolicy` where they need to be able to determine the nature of
the query.

For relationships, this also adds traits that define the expected
cardinality of the result references. This helps inform a host whether
they should be using the paged API methods.

Part of #48

Signed-off-by: Tom Cowland <tom@foundry.com>