[Docs] Add Jupyter notebook covering version workflows
Adds a Jupyter notebook that demonstrates the traits and specifications
that can be used to introspect entity versioning information and
relationships.

It includes a CI task that executes and renders the notebook so we can
ensure that the included code works.

We opted not to `!python -m pip install` dependecies in the notebook
itself to avoid polluting the parent Python environment. The included
`examples/resources/requirements.txt` installs the needed
pre-requisites. Opted to share one requirements file for all notebooks
for now to simplify setup.

Signed-off-by: Tom Cowland <tom@foundry.com>