[Docs] Add entityTraits notebook
Part of OpenAssetIO/OpenAssetIO#31. This notebook was originally used to
sketch out the design of the `entityTraits` API method, elicit
discussion, and come to a shared understanding of the feature as a team.

Being an executable document, the notebook now functions as tested
documentation and an e2e test suite for the feature, utilising the
OpenAssetIO ecosystem beyond the core `entityTraits` API method
(i.e. convenience signatures, BAL, `managementPolicy`).

Signed-off-by: David Feltell <david.feltell@foundry.com>