[Tests] Add Specification tests
Asserts they contain a single usage trait. This is required for runtime
introspection of the puropse of the spec.

Asserts Specification traits constructible, as we were implicitly
testing this as part of checking the trait set contained a usage trait
(as the traitgen uses their attr to build the set), but that was relying
on an implementation details. This requires that MediaCreation
doesn't reference any other packages (which it doesn't at present).

Asserts relationship specication names follow guidelines.

Signed-off-by: Tom Cowland <tom@foundry.com>