[Docs] Add generic republish notebook
Part of OpenAssetIO/OpenAssetIO#31. This notebook came as an appendix
to illustrating usage of `entityTraits` in the original design notebook
for that API method (see previous commit).

However, during review it became apparent that a generic resolve and
republish workflow is a distinct use-case, and not specific to
`entityTraits`. So it was moved to its own notebook.

Signed-off-by: David Feltell <david.feltell@foundry.com>