[Docs] Trait versioning design sketch
Closes #88. As due diligence before going ahead with an implementation,
we need to sketch out what a versioned trait package would look like,
and how hosts and managers would interact with it.

So add a contrived trait package with two versions. The "generated"
view classes have been manually created to reflect what the `traitgen`
output should be, once implemented.

These view classes are then used in a tutorial notebook that describes
how one would go about working with versioned traits.

This work may be abandoned or adapted in favour of more concrete
documentation as the feature is developed.

Signed-off-by: David Feltell <david.feltell@foundry.com>