[Docs] Add string substitution notebook
Closes #91. A gaping hole in OpenAssetIO standardisation is how strings
with embedded substitution tokens should be communicated between a host
and a manager. Hosts and managers must agree on the syntax of
substitution tokens, what tokens are available, and the meaning of those
tokens.

So add a notebook explaining the OpenAssetIO approach to solving this,
including example usage of the `subtitute` utility function.

Signed-off-by: David Feltell <david.feltell@foundry.com>