[CI] Use traitgen 1.0.0a9 for C++ tests
Older versions of `traitgen` used the old namespace location for types
that were later moved. Originally we had aliases for backward
compatibility. But support for deprecated aliases was dropped in
OpenAssetIO/OpenAssetIO#1311.

We bumped the version of `traitgen` used in `pyproject.toml` as part
of the previous release, but didn't update the independent installation
in CI (we have to install independently for the C++ tests).
Because of this, the C++ tests in CI fail.

So bump `traitgen` in the C++ CI to the same version that is listed in
`pyproject.toml`.

Signed-off-by: David Feltell <david.feltell@foundry.com>