[Build] Drop support for CY22
Part of OpenAssetIO/OpenAssetIO#1351.

Disallow Python 3.9 or below for the next release. Technically this
isn't (yet) necessary, but since we're no longer testing on Python 3.9,
any future use of 3.10+ specific features that are not compatible with
3.9 will not get flagged. So play it safe and set the minimum Python
version to 3.10.

Remove libstdc++ ABI toggle, which was only used in C++ tests to ensure
compatibility with the provided OpenAssetIO build.

Signed-off-by: David Feltell <david.feltell@foundry.com>