[Traits] Add initial ui and uiPolicy traits
Part of OpenAssetIO/OpenAssetIO#1302.

Add initial tranche of traits that were found to be useful when
prototyping UI delegation, and update some existing traits. In
particular:

Add traits for sets of images. An image collection could be a manifest,
directory, archive file, templated path, etc, containing one or more
images. An image sequence specialises image collection with a frame
range, implying the images in the collection are ordered. E.g. an image
collection with a MIME type of "inode/directory" refers to a directory
containing images.

Expand on MIME type usage in LocatableContent trait. In particular,
support for multiple MIME types in a single field; and sources of common
MIME types.

Signed-off-by: David Feltell <david.feltell@foundry.com>