[Docs] Update README CMake instructions to include Python by default
It turns out in practice that users usually want both the C++ and
Python component when generating the trait library using CMake. There
was no mention that the Python component must be enabled using a CMake
variable.

So add a small tweak to the README so that the default instructions
enable the Python component.

Signed-off-by: David Feltell <david.feltell@foundry.com>