[CI] Fail if any notebook fails
Part of OpenAssetIO/OpenAssetIO#1302.

The CI job that checks if notebooks can be executed was expected to fail
until upstream OpenAssetIO changes were available, however the job
succeeded. Upon checking the logs we can see the failure of the new
notebook, but the job finished without error. It seems that `find -exec`
doesn't error.

Signed-off-by: David Feltell <david.feltell@foundry.com>